redefine zmessage and grpc upgrade
