+++ .gitignore
@@ -1,0 +1,84 @@
+# This file should only ignore things that are generated during a `x.py` build,
+# generated by common IDEs, and optional files controlled by the user that
+# affect the build (such as config.toml).
+# In particular, things like `mir_dump` should not be listed here; they are only
+# created during manual debugging and many people like to clean up instead of
+# having git ignore such leftovers. You can use `.git/info/exclude` to
+# configure your local ignore list.
+
+## File system
+.DS_Store
+desktop.ini
+
+## Editor
+*.swp
+*.swo
+Session.vim
+.cproject
+.idea
+*.iml
+.vscode
+.project
+.favorites.json
+.settings/
+.vs/
+
+## Tool
+.valgrindrc
+.cargo
+# Included because it is part of the test case
+!/tests/run-make/thumb-none-qemu/example/.cargo
+
+## Configuration
+/config.toml
+/Makefile
+config.mk
+config.stamp
+no_llvm_build
+
+## Build
+/dl/
+/doc/
+/inst/
+/llvm/
+/mingw-build/
+build/
+!/compiler/rustc_mir_build/src/build/
+/build-rust-analyzer/
+/dist/
+/unicode-downloads
+/target
+/src/bootstrap/target
+/src/tools/x/target
+# Created by default with `src/ci/docker/run.sh`
+/obj/
+
+## Temporary files
+*~
+\#*
+\#*\#
+.#*
+
+## Tags
+tags
+tags.*
+TAGS
+TAGS.*
+
+## Python
+__pycache__/
+*.py[cod]
+*$py.class
+
+## Node
+node_modules
+package-lock.json
+package.json
+
+## Rustdoc GUI tests
+tests/rustdoc-gui/src/**.lock
+
+# Before adding new lines, see the comment at the top.
+
+/db
+/log
