+++ Cargo.toml
@@ -7,19 +7,20 @@ edition = "2021"
 
 [dependencies]
 zchronod-logger = { version = "0.1.0", path = "./logger" }
-process = {version = "0.1.0",path ="./process"}
+process = { version = "0.1.0", path = "./process" }
 clap = "2.33.3"
-tokio = {  version = "1.35.1", features = ["full"] }
+tokio = { version = "1.35.1", features = ["full"] }
 matches = "0.1.10"
-api ={version ="0.1.0", path="./api"}
-proto ={version = "0.1.0",path = "./proto"}
-storage = {version = "0.1.0", path="./storage"}
+api = { version = "0.1.0", path = "./api" }
+proto = { version = "0.1.0", path = "./proto" }
+storage = { version = "0.1.0", path = "./storage" }
+client = { version = "0.1.0", path = "./client" }
 
 [workspace]
-members = ["proto"]
+members = ["proto", "client"]
 
 [workspace.dependencies]
-env_logger = "0.9.3"
+env_logger = "0.11.3"
 lazy_static = "1.4.0"
 log = "0.4.20"
 log4rs = "1.0"
