+++ client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "proto"
+name = "client"
 version = "0.1.0"
 edition.workspace = true
 license.workspace = true
@@ -10,8 +10,7 @@ authors.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tonic-build = "0.10.2"
-prost = "0.12.3"
-tonic = "0.10.2"
-serde = { version = "1.0.196", features = ["derive"] }
-tokio = { version = "1.0.0", features = ["rt", "rt-multi-thread", "macros"] }
+tokio = "1.36.0"
+tonic = "0.11.0"
+proto = { version = "0.1.0", path = "../proto" }
+
