+++ process/src/lib.rs
@@ -15,6 +15,7 @@ use libp2p::{
     noise,
     PeerId, swarm::{NetworkBehaviour, SwarmEvent}, Swarm, tcp, yamux,
 };
+use libp2p::dns::ResolveErrorKind::Proto;
 use log::{error, info};
 use prost::Message as m1;
 use serde::{Deserialize, Serialize};
@@ -25,11 +26,12 @@ use api::{CONTEXT, NetworkInterface, Node};
 use chronod::clock::{Clock, VlcMeta, VlcMsg, ZMessage};
 use network::{GossipServer, RpcServer};
 use proto::zchronod::Event;
-use proto::zchronod::zchronod_server::Zchronod;
+use proto::zmessage::ZType;
+use proto::zmessage::ZType::Rng;
 use storage::ZchronodDb;
 
 pub struct ZchronodServer {
-    gossip_send: tokio::sync::mpsc::Sender<ZMessage>,
+    gossip_send: tokio::sync::mpsc::Sender<proto::zmessage::ZMessage>,
     node_address: String,
     // todo add config as node_config
     inner: Arc<RwLock<CoreZchronod>>,
@@ -38,23 +40,26 @@ pub struct ZchronodServer {
 
 impl ZchronodServer {
     // assume rpc is a new from
-    fn handle_rpc_msg(&self, x: Event) {
+    fn handle_rpc_msg(&self, x: proto::zmessage::ZMessage) {
         println!("receive from rpc {:?}", x);
 
         // construct publish event to gossip
-        self.inner.write().unwrap().count += 1;
-        println!("current inner count is  {}", self.inner.read().unwrap().count);
-        println!("current clock value is  {}， should+1", self.inner.read().unwrap().clock.get_value());
-        self.inner.write().unwrap().clock.inc();
+
+        let mut inner = self.inner.write().unwrap();
+        inner.count += 1;
+        println!("current inner count is  {}", inner.count);
+        println!("current Clock value is  {}， should+1", inner.clock.get_value());
+        inner.clock.inc();
+
 
         // construct z_message
         let mut event_bytes = Vec::new();
         x.encode(&mut event_bytes).unwrap();
 
         let clock_msg = Clock {
-            id: self.inner.read().unwrap().clock.id.clone(),
-            value: self.inner.read().unwrap().clock.value,
-            ancestors: vec![self.inner.read().unwrap().clock.clone()],
+            id: inner.clock.id.clone(),
+            value: inner.clock.value,
+            ancestors: vec![inner.clock.clone()],
         };
         let vlc_request_message = VlcMeta {
             clock_state: Some(clock_msg),
@@ -71,48 +76,24 @@ impl ZchronodServer {
         let mut z_mes_bytes = Vec::new();
         vlc_msg.encode(&mut z_mes_bytes).unwrap();
 
-        let z_message = ZMessage {
-            r#type: "vlc".to_string(),
-            msg_meta: z_mes_bytes,
+        let z_message = proto::zmessage::ZMessage {
+            version: 0,
+            r#type: Rng as i32,
+            public_key: vec![],
+            data: z_mes_bytes,
+            signature: vec![],
         };
 
         println!("z-message construct ok, save to db, send to gossip");
-        //   if x.kind == 301 {}
 
-        self.distribute_event_msg_to_db(x);
+        //self.distribute_event_msg_to_db(x);
         let rt = Runtime::new();
         rt.unwrap().block_on(self.gossip_send.send(
             z_message)).expect("failed to send to gossip");
     }
 
-    fn distribute_event_msg_to_db(&self, e: Event) {
-        println!("distribute rpc msg here");
-        if e.kind == 301 {
-            println!("receive kind 301 poll");
-            info!("receive kind 301 poll");
-            match self.z_db.write().unwrap().poll_write(self.construct_poll_event_key(e.clone()), e.clone()) {
-                Ok(_) => {}
-                Err(_) => { return; }
-            }
-        }
-
-        if e.kind == 309 {
-            println!("receive kind 309 vote");
-            info!("receive kind 309 vote");
-            match self.z_db.write().unwrap().vote_write(e.clone()) {
-                Ok(_) => {}
-                Err(_) => { return; }
-            }
-        }
-
-        match self.z_db.write().unwrap().event_write(e.clone()) {
-            Ok(_) => {}
-            Err(_) => {
-                error!("event id duplicated");
-                println!("event id duplicated");
-                return;
-            }
-        }
+    fn distribute_event_msg_to_db(&self, e: proto::zmessage::ZMessage) {
+        todo!()
     }
 
     fn construct_poll_event_key(&self, e: Event) -> String {
@@ -145,19 +126,18 @@ impl ZchronodServer {
         self.inner.write().unwrap().clock.merge(&vec![&clock_state]);
         //self.inner.write().unwrap().clock.inc();
         println!("current inner count is  {}", self.inner.read().unwrap().count);
-        println!("current clock value is  {}， should+1", self.inner.read().unwrap().clock.get_value());
+        println!("current clock value is  {}, should+1", self.inner.read().unwrap().clock.get_value());
 
-        let e = Event::decode(Bytes::from(vlc_meta_instance.event_meta)).unwrap();
-        self.distribute_event_msg_to_db(e);
-        // self.inner.write().unwrap().clock.inc();
+        //let e = Event::decode(Bytes::from(vlc_meta_instance.event_meta)).unwrap();
+        //self.distribute_event_msg_to_db(e);;
     }
     pub fn handle_gossip_msg(&self, z_msg_bytes: Vec<u8>) {
         println!("handle gossip msg");
-        let z_message = ZMessage::decode(Bytes::from(z_msg_bytes)).unwrap();
-        match z_message.r#type.as_str() {
-            "vlc" => {
+        let z_message = proto::zmessage::ZMessage::decode(Bytes::from(z_msg_bytes)).unwrap();
+        match z_message.r#type {
+            0 => {
                 println!("handle vlc msg");
-                let vlc_msg = VlcMsg::decode(Bytes::from(z_message.msg_meta)).unwrap();
+                let vlc_msg = VlcMsg::decode(Bytes::from(z_message.data)).unwrap();
                 match vlc_msg.r#type.as_str() {
                     "request" => {
                         println!("handle vlc_request msg");
@@ -175,20 +155,6 @@ impl ZchronodServer {
                 println!("unknown z_message type");
             }
         }
-
-        // let clock_other = Clock::decode(Bytes::from(clock_bytes)).unwrap();
-        //  println!("receive from gossip [{}]", &clock_other.id);
-        //  match self.inner.read().unwrap().clock.partial_cmp(&clock_other) {
-        //      Some(Ordering::Greater) => return,
-        //      _ => {
-        //          println!("need merge");
-        //          self.inner.write().unwrap().clock.merge(&vec![&clock_other]);
-        //      }
-        //  }
-        //  self.inner.write().unwrap().clock.inc();
-        //  self.inner.write().unwrap().count += 1;
-        //  println!("current inner count is  {}", self.inner.read().unwrap().count);
-        //  println!("current clock value is  {}， should+1", self.inner.read().unwrap().clock.get_value());
     }
 }
 
@@ -238,21 +204,18 @@ pub fn init_chrono_node(config: &str) {
     // network::set().expect("TODO: panic message");
 }
 
-fn run(mut gossip: GossipServer<ZMessage>, db: Arc<RwLock<ZchronodDb>>, rpc: RpcServer, id: String) {
-    println!("run");
-
+fn run(mut gossip: GossipServer<proto::zmessage::ZMessage>, db: Arc<RwLock<ZchronodDb>>, rpc: RpcServer, id: String) {
     let sender = gossip.send.clone();
+    let sender_copy = gossip.send.clone();
     let consensus = Arc::new(chronod::init());
     let db_rpc_service = Arc::clone(&db);
     //let consensus_clone = Arc::clone(&consensus);
-    rpc.run(gossip.send.clone(), consensus.receive(), db_rpc_service).expect("failed to run rpc");
+    rpc.run(consensus.send.clone(), db_rpc_service).expect("failed to run rpc");
     let (gossip_send, gossip_recv) = mpsc::channel::<(PeerId, Message)>();
     gossip.register_receive(gossip_send);
 
-    let inner = Arc::new(RwLock::new(CoreZchronod { count: 0, clock: chronod::Clock::new(id) }));
-    let sender_copy = gossip.send.clone();
-    let inner_c = Arc::clone(&inner);
-
+    let inner = Arc::new(RwLock::new(CoreZchronod { count: 0, clock: Clock::new(id) }));
+    let inner_clone = Arc::clone(&inner);
 
     let db_c = Arc::clone(&db);
     thread::spawn(move || {
@@ -270,7 +233,6 @@ fn run(mut gossip: GossipServer<ZMessage>, db: Arc<RwLock<ZchronodDb>>, rpc: Rpc
         }
     });
 
-    // tokio::spawn(async move {gossip.start();});
     thread::spawn(move || {
         let rt = Runtime::new().unwrap();
         rt.block_on(async { gossip.start().await })
@@ -283,7 +245,7 @@ fn run(mut gossip: GossipServer<ZMessage>, db: Arc<RwLock<ZchronodDb>>, rpc: Rpc
             let handle1 = ZchronodServer {
                 gossip_send: sender.clone(),
                 node_address: "".to_string(),
-                inner: inner_c.clone(),
+                inner: inner_clone.clone(),
                 z_db: db_c.clone(),
             };
