+++ proto/src/build.rs
@@ -1,10 +1,10 @@
-
 use std::io::Result;
 
 // used to build proto
 pub fn set() -> Result<()> {
     tonic_build::configure()
         .out_dir("proto/src")
-        .compile(&["src/zchronod.proto"], &["proto/"])?;
+        .type_attribute(".", "#[derive(serde::Serialize,serde::Deserialize)]")
+        .compile(&["src/zchronod.proto", "src/zmessage.proto", "src/msg.proto"], &["proto/"])?;
     Ok(())
 }
