+++ proto/src/lib.rs
@@ -2,3 +2,5 @@ pub mod build;
 
 pub mod zchronod;
 pub mod msg;
+
+pub mod zmessage;
